Updated form validation logic to get matlab root value.
